% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_parla_clarin_xml_file.R
\name{read_parla_clarin_xml_file}
\alias{read_parla_clarin_xml_file}
\title{Read a Parla-Clarin XML file}
\usage{
read_parla_clarin_xml_file(x, ...)
}
\arguments{
\item{x}{a file path to a Parla-Clarin XML file}

\item{...}{further arguments to \code{xml2::read_xml()}.}
}
\description{
This function reads in a Parla-Clarin XML file and converts it to an
R list object with class \code{parla_clarin_xml}.
}
\details{
For details on the Parla-Clarin XML schema, see:
\url{https://github.com/clarin-eric/parla-clarin}
}
